+++ src/Jadob/Bridge/Twig/Extension/PathExtension.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jadob\TwigBridge\Twig\Extension;
+namespace Jadob\Bridge\Twig\Extension;
 
 use Jadob\Router\Router;